+++ test/fixtures/project_unused_decl/src/lib.zig
@@ -1,3 +1,3 @@
-pub const PublicApi = @import("public_api.zig");
+pub const PublicApi = @import("./public_api.zig");
 
 pub fn rootEntry() void {}