+++ test/fixtures/project_unused_decl/usingnamespace_pkg/src/lib.zig
@@ -1,0 +1,1 @@
+pub usingnamespace @import("./public_api.zig");